Eliminate duplicated data type definitions in makedoc
makedoc.c defined an independent enum for DT_... muttrc variable
types, for which when adding a new type to init.h one had to think
of adding a new enum value and entry to types[] as well.

Split the DT_... defines into a separate datatypes.h and include
that and use them to define the types[] elements.

If a new DT_... is used in init.h then buff2type() now bails out
if it can't find it in types[].

Also, handle each data type explicitly in pretty_default() and
bail out if an unhandled type is encountered.