de.po: proper translation for term collapse
The intention of the term "collapse" is something like opening and
closing an item. The initial translation which was added two decades
ago means something like "breaks by weakness", which does not fit in
the context of mutt.

Use a term which means literally "opening and closing an item, like a
book".

Signed-off-by: Olaf Hering <olaf@aepfle.de>