Add mutt_oauth2.py command line parameters
This extends mutt_oauth2.py with more command line parameters to avoid
the necessity of 1. patching the script and 2. answering interactive
questions. The documentation is also updated to reflect this change.

Anybody prefering the old way is of course still free to hard-code the
gpg and other IDs in the source code. If the command line options for
the --authorize mode are not supplied, the script falls back to the
interactive mode as before.