Abort macros on unexpected prompts
During the interpretation of a macro, if we get an input key that does
not match one of the possible choices of a prompt, it probably means
that the macro did not expect that prompt to appear. The safest way to
deal with such a case is to abort the macro.
Define a mutt_flush_macro_on_error function for that.