Abort execution of <func> when in an incompatible context
When a function is encountered in a macro using the syntax "<func>" but
that it cannot be invoked in the context, the main cause is most
probably because the macro didn’t perform what was expected.
Instead of then interpreting "<func>" as the sequence of all its keys
(which could perform really undesirable actions), simply abort.